+++ history.md
@@ -2,7 +2,6 @@
 
 
 # Features planned (in random order)
-- []   Add more Update/Replace Tests
 - []   Front-end for Replace API in folder view
 - []   Realtime Files API
 - []   Performance upgrade xmp/tiff files
@@ -29,6 +28,7 @@
 - Done some IStorage refactorings, but not complete yet
 - Next/Prev links are now served by backend code to avoid when javascript is not loaded, your selection is reset
 - Add Collections to DetailView model
+- [x]   Add more Update/Replace Tests
 
 # version 0.1.4 - 2019-03-01
 - fix issue where login fails results in a error 500